also move set-tool actions into correct modules
