+++ frontend/javascripts/oxalis/api/api_latest.js
@@ -83,7 +83,8 @@ import {
   setTreeGroupsAction,
 } from "oxalis/model/actions/skeletontracing_actions";
 import { setPositionAction, setRotationAction } from "oxalis/model/actions/flycam_actions";
-import { setToolAction, refreshIsosurfacesAction } from "oxalis/model/actions/annotation_actions";
+import { refreshIsosurfacesAction } from "oxalis/model/actions/annotation_actions";
+import { setToolAction } from "oxalis/model/actions/ui_actions";
 import {
   updateUserSettingAction,
   updateDatasetSettingAction,
