+++ frontend/javascripts/oxalis/view/action-bar/toolbar_view.js
@@ -15,7 +15,7 @@ import { document } from "libs/window";
 import { enforceVolumeTracing } from "oxalis/model/accessors/volumetracing_accessor";
 import { getRenderableResolutionForSegmentation } from "oxalis/model/accessors/dataset_accessor";
 import { createCellAction } from "oxalis/model/actions/volumetracing_actions";
-import { setToolAction } from "oxalis/model/actions/annotation_actions";
+import { setToolAction } from "oxalis/model/actions/ui_actions";
 import { updateUserSettingAction } from "oxalis/model/actions/settings_actions";
 import { usePrevious, useKeyPress } from "libs/react_hooks";
 import ButtonComponent from "oxalis/view/components/button_component";
